Only use one `PerBuildState` per `cquery`/`query`
Summary:
I noticed when looking at chrome traces that all the pipelines had multiple entries - they had entries at the top level (representing the `PerBuildState` that is generated at the beginning of the query) and then they also had entries in the middle of `creating_precomputed_target_universe`. This suggests that there are two `PerBuildState`s hanging around.

Thanks to the debugger this was easy to find - the `resolveTargetSpecs` method on `Parser` was taking a `ParsingContext` and then creating it's own `PerBuildState`. This is silly when we already have one of those objects hanging around.

This diff makes it so that `Parser` has a method that takes an external `PerBuildState` rather than creating it's own and then makes `cquery`/`query` use that method.

Reviewed By: bobyangyf

shipit-source-id: e05880dc4844bf4cbd43ef66399f3f2c5549c14a