Remove `ParsingContext` from `TargetUniverse.resolveTargetSpecs`
Summary:
Now that `LegacyQueryUniverse` doesn't use it (and since `PrecomputedTargetUniverse` never used it) this parameter is completely unused by all implementations of `TargetUniverse`. Therefore we can get rid of it.

(This also gives `TargetUniverse` a much better API, as we aren't passing a random object in from above just because one implementation needed it)

Reviewed By: stepancheg

shipit-source-id: 8ce3f6fb3659f8806d3e86642bd1481af14aa319