Move Unconfigured graph maintenance to another class
Summary: Most of the code in `UnconfiguredQueryEnvironment` was about maintaining the graph of `UnconfiguredTargetNode`s, and it was starting to get unweildy. I needed to make a pretty major change anyway (the move away from explicit deps caching to caching the traversal of all attributes) so this seemed like a good time to make the swap.

Reviewed By: stepancheg

shipit-source-id: 5b2c65ab8f1df205c651c2976f9ad0b9ad43c2c6