Split `UnconfiguredCxxGenruleFilterAndTargetsMacro` into multiple types
Summary:
Unfortunately the `typeCoercer` infrastructure requires that there is a single type coercer per type, and _technically_ we have 5+ for `UnconfiguredCxxGenruleFilterAndTargetsMacro` (one for each subgroup).

We would have to do some serious restructuring to get the `TypeCoercer` infra to realize that these are all coercers for different types. Lets do it the easy way, following what we have for the configured macros - tons of classes.

Reviewed By: stepancheg

shipit-source-id: 0eef4a25f05358e675327e704437361ec67eb0dc