Make uquery capable of calculating parse deps of a node
Summary: `cquery` didn't have to deal with this since `TargetNode` did this for us, but we're in the wild west with `uquery`. This diff makes use of the `traverseUnconfigured` method I added everywhere in the last two diffs to go through the object hierarchy and find all the build targets, so we know what to parse.

Reviewed By: stepancheg

shipit-source-id: 801e708f27925a11533438a6adeb7d4d44c69229