cxx.toolchain_target must be a target
Summary:
`getMaybeUnconfiguredBuildTarget` returns `Optional.empty()` when buckconfig value is not a valid target.

So specifying `cxx.toolchain_target=/rrrr` will result in Buck ignoring this property. We want to explicitly crash instead.

`getMaybeUnconfiguredBuildTarget` function is used when the same buckconfig value can be both target or path, which is not the case for `cxx.toolchain_target`.

Reviewed By: milend

shipit-source-id: c4d16f3322155f263b810f30377fcf5ebfe13e5e