Add IsolatedShellStep
Summary:
The objective of adding `IsolatedShellStep` is to migrate `ZipAlignStep` to an `IsolatedStep`. However, `ZipAlignStep` extends `ShellStep`, so it can't also extend `IsolatedStep`.

**Options considered:**
1. Copy all the code from `ShellStep` into `ZipAlignStep`
2. Make `ShellStep` extend `IsolatedStep`
3. **[This diff]** Make an IsolatedShellStep that does what ShellStep currently does. `ShellStep` will hold an instance of `IsolatedShellStep` and delegate all calls to `IsolatedShellStep` (with the exceptions of `getShellCommand` and `getDescription`, which actually need `StepExecutionContext`)

Chose option #3.

Why not option #1: Requires maintaining the same code in two different places. Probably okay if the code is simple, but this code is non-trivial and error-prone to maintain separately.

Why not option #2: This approach requires changing all the other subclasses of `ShellStep` to implement the new `IsolatedStep` methods. Choosing #3 avoids having to change a bunch of other subclasses

**Implementation**

*IsolatedShellStep*

`IsolatedShellStep` is basically a copy of the old `ShellStep` with a few the following differences:
* Instead of changing `execute()` directly to `executeIsolatedStep()`, I added a new `executeCommand()` method that both `ShellStep` and `IsolatedShellStep` can use. This is so code can be shared between the two after resolving their respective args into list of string
* `executeCommand()` has a lot more arguments than `execute()` and `executeIsolatedSteps()` to propagate the overridden behavior from the subclasses properly. At some point this class should probably be refactored to have a bunch of `final` member vars that holds all these parameters. That's probably easier to maintain and understand. The current code flow is hard to follow.
* Method visibilities changed to `public` so can be called by `ShellStep`
* Added javadocs

`IsolatedShellStep` is added underneath a new `common` dir in `buck/step/isolatedsteps`. I think `shell` rather than `common` could be a good name too. Any preferences?

*ShellStep*
* As aforementioned, `ShellStep` now holds an `IsolatedShellStep` and calls methods on `IsolatedShellStep` to avoid code duplication.
* Note that the `IsolatedShellStep` it holds throws ann exception if `getShellCommandInternal(IsolatedExecutionContext context) ` is called. This is because if `ShellStep` is being used, `IsolatedShellStep`'s `getShellCommand()`, which in turn calls `getShellCommandInternal()`, should never be called due to context mismatch.

 ---

With `IsolatedShellStep` in place, `ZipAlignStep` can be changed to extend `IsolatedShellStep` and be migrated to an `IsolatedStep` in a follow-up diff

Reviewed By: mykola-semko

shipit-source-id: 11a9f3727b840e8ef6890cbea3ad493279a28b36