Refactor parsing of bind symbols
Summary:
We want the ability to extract bind symbols from two Mach-O sections (they have the same format), so here we just extract a common builder that will be used by both.

In addition, we remove the wrapping of the results in an `Optional` as that was not useful in practice.

Reviewed By: selat

shipit-source-id: 00a7670c1e564f1608a79314c98fe33a27202230