Remove `TargetConfiguration` as input to `TargetUniverse`
Summary:
When I first set up the `TargetUniverse` interface all I knew was that `PrecomputedTargetUniverse` couldn't hold on to the `parser` so I replaced all parser methods with `TargetUniverse` methods. During that process, I used the same parameters because I didn't know any better.

Well now I do, and I have a better understanding of what each method _does_. The current setup meant we were simply ignoring the `TargetConfiguration` in a bunch of `PrecomputedTargetUniverse` methods, when in reality we should have given the `TargetConfiguration` to `LegacyQueryUniverse` to begin with.

(This also means we can remove `TargetConfiguration` from a bunch of classes that had to hold onto it before)

Reviewed By: stepancheg

shipit-source-id: 2221556dc5680fd2facc18bb4049bd4b2857893d