Make SwiftCompile rule support input-based rulekey
Summary:
The SwiftCompile rule did not support input-based rulekey before, here we opt-in the rule, so that it avoids unnecessary recompilation of Swift dependents if the public API of a module has not changed.

For context, Swift generates `.swiftmodule` files which define the API exposed by a particular module. Dependent libraries use those `.swiftmodule` files to type-check the Swift code being compiled.

Previously, we were passing the Swift output dir which contained the `.swiftmodule`, in addition to the object file and generated `-Swift.h` headers. This meant that changes to the private implementation of the Swift module triggered recompilation of dependents.

In this diff, we only expose the `.swiftmodule` to dependents, so any private changes do not trigger unnecessary recompilation.

NOTE: This is a re-import of the Airbnb [pull request](https://github.com/airbnb/buck/pull/23) with some small tweaks.

Reviewed By: cjhopman

shipit-source-id: 397bf7e12aca7ca93aba1acc8eaa0536ac552b39