Add java11_for_tests config option
Summary:
This is needed to enable us to run some tests with java 11 and others on a different version (e.g. java 8). It was suggested in the switchover plan: https://fb.quip.com/cua4APsYfbM4

If a test specifies that it uses java 11, then it will be run with java 11. Otherwise it will just use the tool specified at `java_for_tests`.

Reviewed By: mykola-semko

shipit-source-id: e754e201837581c45e77e22941daf089edf974f1