Add new page for overview of Buck Query Language
Summary:
I want to add docs for `cquery` and `uquery`, but a large portion of the `query` docs are a high level overview of the query language. Since this overview will be the same for all three query commands, split it into it's own page (and link to it) rather than copy-paste the content.

NOTE: Since this section was originally written to all be in one page, I had to make a few links from the `query` docs back to the `query_language` page, and had to modify the wording of some of the paragraphs to specify that things like "which functions are available" are different depending on context. I left the examples all referencing `query` though, for simplicity's sake.

Reviewed By: cjhopman

shipit-source-id: c8c37e4ad21b9cfdfd2dee9863c227e090831944