Don't create a class ABI for AndroidPrebuiltAar
Summary: AndroidPrebuiltAarDescription is responsible for creating a number of different flavors, including the `AndroidPrebuiltAar` rule itself, and also the `#prebuilt_jar` flavor. Due to this code: https://fburl.com/diffusion/gnsvqnj2, we were creating an identical class-abi for both of these flavors (`rule#class-abi` and `rule#prebuilt_jar,class-abi`). The `rule#class-abi` was the one that shouldn't exist (it was the class ABI for the `prebuilt-jar` flavor. We just remove this rule entirely.

Reviewed By: asp2insp

shipit-source-id: 310d5f1cc678925a7d79fef3627227ddfe37e318