buck/rust: add experimental support for --extern-location
Summary:
This corresponds to experimental upstream Rust PR https://github.com/rust-lang/rust/pull/72603

This is not intended to be a stable feature; it could change as the
corresponding rustc functionality does.

`--extern` is the parameter Buck passes rustc for each direct dependency of a
crate, so that it can load its definitions, etc. I recently added a
-Wunused-crate-dependencies diagnostic which warns if there are `--extern`
crates which are never referenced, indicating unused dependencies. This works
well but there's not really enough information to really identify where the
dependency comes from.

This adds `--extern-location <name>=json:{"target":<buck
target>,"dep":<dependency target>}` which allows a tool to precisely identify
and fix the excess dependency.

This is enabled with the `rust.extern_locations` config option.

Reviewed By: cjhopman

shipit-source-id: ab9ec403893bf94faf42739401d0722b9012d14e