Extract `uquery` unconfigured attribute traversal to new class
Summary: This better encapsulates the logic since it isn't really that relevant to the `UnconfiguredTargetGraph` broadly. All the target graph really wants/needs is the ability to say "give me the deps of X" or "give me the inputs to X" - it doesn't need control over how that information is calculated.

Reviewed By: cjhopman

shipit-source-id: 00413530dd6865e631bc24fcf21ede4490ffcd9f