Use compile-time classpath for Java 8 desugar
Summary:
Java 8 desugar requires any default interfaces that were declared to be on the classpath. By using the classpath that was used for compilation, we guarantee that we provide all such interfaces to d8.

This is much more efficient than calculating all of the transitive classpaths, which can be HUGE, and additionally this classpath has already been compiled so costs us nothing.

Reviewed By: brianduff

shipit-source-id: a96618ff298449cc159266a28042947cb892be01