Add getCompileTimeClasspathSourcePaths to HasClasspathEntries
Summary:
This gives us access to the classpath that was used when building the java library.

Please feels free to correct any of the implementations, I wasn't 100% sure about all of them.

This will be used in the next diff because we want to pass the same classpath used for compilation to d8 when doing Java 8 desugaring.

Reviewed By: brianduff

shipit-source-id: 75498eb094542bc1677d1992b39018710a728068