Enable opt-in for remote logging via Scribe and Scribble
Summary:
Adds a new property, `use_scribed` to the `log` field of the the .buckconfig. This field will be responsible for allowing users to asynchronously offload the logging step to a remote process which will upload the logs to Scuba.

The field is being introduced now to provide an easy way to test the asynchronous logging to scribed and scribble.

As of now the feature is opt in only and the field is given a false default value. Making the value true will not affect the build in any way at this time.

Reviewed By: swgillespie

shipit-source-id: eab173c72e268953d34beba8a4e5fe6c8a882957