Remove the string `buck query` from function examples
Summary:
This shows up often in the function documentation, and it's not strictly necessary. The documentation is generally referring to broad functionality. These functions operate the same way between every query, the only differences are things like output format. The query string `kind('java_library', //...)` is just as valid an example for `cquery` as it is for `query`.

Therefore, remove the string `buck query` so that these examples can be used in documentation for other query pages where they are still relevant.

Reviewed By: stepancheg

shipit-source-id: 62cbe03a2d41ab681b3b1150d35ef1d1e90ce618