Introduce 'hit action cache' strategy result for AC hits
Summary:
This commit augments the RemoteExecutionStrategy to report 'hit action cache' as the strategy result for actions that hit the action cache. Whether or not a remote action hit the action cache is stored in the `RemoteBuildRuleExecutionEvent` event fired upon receipt of the action result.

`CriticalPathEventListener` uses this to avoid calling the `onBuildRuleCompletedExecution` critical path callback for remote actions that hit the action cache.

Reviewed By: cjhopman

shipit-source-id: de87dcadbf9c57676d7f6f1a66b62f8891c67fd0