Make `ImplicitInputsInferringDescription` take the raw attributes
Summary:
The constructor arg doesn't always exist in every situation in which we want to use implicit inputs. Notably, `uquery` (which operates on an unconfigured graph) has no concept of a `ConstructorArg` (which only gets created when generating configured target nodes).

This diff makes the interface use the raw attributes, so it can be better reused by `uquery`.

Reviewed By: bobyangyf

shipit-source-id: 307ed522643d7f31e3b1271aa9bdd188c476598c