Explicitly add cxx toolchain dependencies to apple_ rules (attempt 2)
Summary:
This is a second attempt to land D21539304, which was reverted.

This code:

```
getAppleCxxPlatformsFlavorDomain(toolchainProvider, buildTarget.getTargetConfiguration())
```

returns empty list and does not have a chance to add dependency on toolchain rules.

That may happen when we use `apple_library` rules, but only configure `cxx_toolchain` rule.

This diff explicitly add dependency on C++ toolchain for all apple rules same way we do for `cxx_library` rule.

Reviewed By: milend

shipit-source-id: a0a950abff8b07aef564484e39571b4c8dc6ad1d