Make `ClasspathFunction` ignorant of `GraphEnhancementQueryEnvironment`
Summary:
Right now the first thing this function does is assert that the environment passed in is a `GraphEnhancementQueryEnvironment`. This is a problem because `GraphEnhancementQueryEnvironment` only works with `ConfiguredQueryTarget` so we can't use any unconfigured targets with it at all.

There are two ways around this:

* Add the relevant functions to `QueryEnvironment`, then throw `IllegalStateException`s in any environments that don't support the `classpath` function. This is the strategy that most query functions take, but it leads to a pretty bloated `QueryEnvironment` interface (which I suspect is why we avoided this strategy in the first place)
* Use a per-function interface and require environments implement this interface

This diff goes with the second option, with the hope that we can apply this strategy to other functions.

Reviewed By: stepancheg

shipit-source-id: 9680eb6fd7a5950611b6378ea4a602907166d2a1