Propagate exceptions during in process dexing so that they can be logged correctly
Summary:
Not using the console to log these exceptions causes them to be clobbered by other exceptions in the console logs, addressed in the next diff. Not consuming the exceptions will cause buck to log them correctly.

This could have worked in the past because errors would pass from D8 -> diagnosticsHandler -> postCompilationFailureToConsole but that doesn't seem to be working for some compilation failures.

Reviewed By: IanChilds

shipit-source-id: 88f627d9c2a8fc367a7dc6cc353ecd93cfdb2e7e