Introduce a #doc flavor for Rust
Summary:
The first step of the Rustdoc project involves adding a `#doc` flavor for Rust buck builds.

Say for example that you had a Rust target at: `//path/to:target`. If we run buck build with a `#doc` flavor, we'd be able to produce HTML and JavaScript webpages which will use rustdoc:
```
buck build //path/to:target#doc
```

Reviewed By: dtolnay

shipit-source-id: bd4caaf24afc85716e5ca5e9e482f881de58275f