Make MiniAapt also check that dependencies of styles are present
Summary:
This attempts to solve a long-standing complaint that resource libraries are able to build without having all the dependencies they need present, since we didn't check whether the styles being used are present. The build would only fail when the full aapt2 run happens, if the dep wasn't present at that point, which is confusing.

Adding attrs to the list of references does change the output R.txt too - I think that's okay but would appreciate anyone who knows more about this chiming in :)

I will have to put this behind a config option if this is agreed as the correct approach, as this is going to uncover a ton of bugs in fbsource unfortunately.

Reviewed By: brianduff

shipit-source-id: 346b551820f383e3fadfe95fd32c865701b7d1d0