Rename of SourcePathResolver method `getRelativePath` to `getCellUnsafeRelPath`
Summary:
aa`SourcePathResolver.getRelativePath` doesn't work properly f or cross cell case. We need to use `getAbsolutePath` and then relativize to `getProjectFilesystem()` if we want to have a cell relative path to work with (preferred way for IsolatedStep convention).

Let's rename a method to a `getCellUnsafeRelPath` that would keep migrator's attention to this pattern :
* getAbsolutePath and then relativize,
* or `getRelativePath` and provide `ProjectFilesystem` as
a parameter

Reviewed By: stepancheg

shipit-source-id: 75348fef2701fc60d4fa8615703d1e2852df5507