Add Thrift Scribe Client to Buck
Summary:
This diff adds an interface and a simple implementation for logging arbitrary messages to the scribed and scribble processes via thrift. The purpose of this follows an ongoing project to eventually offload most of buck's logging to these processes using this interface.

One can create a scribe thrift client and simply pass it along a string for a scribe category and a list of strings containing messages to be sent to the category. The messages are then sent along to the daemon processes who are in charge of sending the data to their respective destination. One can also create their own LogEntry messages directly to have a bit more control over what categories get what message.

A specific example would be using the thrift client to implement a ScubaLogger to upload buck logs via the `perfpipe_buck_builds` scribe category.

Reviewed By: swgillespie

shipit-source-id: c28f8ba1fdba26aa6c850acaf7dffcb9740f3088