Add TopSlowTargetBuilder for keeping track of slow targets
Summary: The simple console today keeps track of the top 10 slow rules for a particular build so that they can be printed out at the end. This is a useful thing to show and so it's a shame that it's limited only to the simple console. This commit splits out and refactors the logic for calculating the top N slowest targets in a build so that it can be re-used by the simple console and other things, with the ultimate goal of getting the top N slow rules for a particular build into Scuba.

Reviewed By: asp2insp

shipit-source-id: 51af5eeb5553ce70e050d916a39621adb611c51d