Add the `fully_qualified_name` computed attribute to `cquery`
Summary:
I thought this was on `query` as well, but apparently this only exists in `buck targets`. For `query` this isn't a big deal though because the key to the output map is the same value as `fully_qualified_name`. For `cquery` this is a problem because that same key is `//foo:foo (//some:platform)`, which makes retrieving the `fully_qualified_name` impossible without string manipulation.

This diff fixes it by explicitly adding the `fully_qualified_name` attribute.

Reviewed By: stepancheg

shipit-source-id: f62a9f59102af105e7d14177dc96837ef47e2d0c