pull native module into global scope of bzl files
Summary:
Allow top level `read_config` and other functions in `.bzl` files.

Make Python DSL and Starlark consistent. Maybe having global `read_config` in not the best idea, but if we really need to do that, we should migrate repositories first, and disable these functions in Python DSL too.

This change is also needed to make `load_symbols` work with the change in the next diff.

Reviewed By: cjhopman

shipit-source-id: 5daeb215e03575c00502b5506aab4e43b2c4caf7