New ScribeLogger Implementation using Remote Thrift Client
Summary:
### Added Remote Thrift Scribe Logger

Added a new implementation of the ScribeLogger using the rpc thrift bindings brought in in previous diffs. Scribe logger strives to provide the same promises by the current ThriftScribeLogger.

* Remove null values from given lines of data

* Always resolves a log() call even if the underlying transport service is closed

* Send filtered lines to a scribe interface to send data to scribe

* Handles runtime errors raised by scribe interface

Since the new implementation offloads the actual logging to a new process there is no threading through Executor services but that change is straightforward should this affect buck's performance.

### Small Changes to the Scribe Client interface
Added the Optional parameter to Scribe Client interface to allow some flexibility of which buckets as it was also promised in the Scribe Logger interface. This is so not every level of the scribe logging has to worry about encoding strings into UTF8 unless even greater flexibility is needed.

The interface now also throws Runtime Errors instead of IO exceptions. I think this was raised in the last diff but I missed it.

#### Note on Runtime Errors
The current Thrift Scribe Logger implementation has the underlying Executor service handle and bubble up runtime errors thrown by the Scribe transport interface. The Scribe Logger interface does not raise errors itself so right now errors are logged using the buck LOG as errors. This might be subject to change should this current method get in the way of figuring out that we might have to use an offline logger or another scribe logger.

Reviewed By: mykola-semko

shipit-source-id: a76c92de004583b9bc953ea5b59e818e84e12516