Simplify pojoizer
Summary:
`BuildFileManifestPojoizer` was used to normalize objects in Python and Starlark.

Now it is only used in Starlark.

Historically it is a bit overengineered, and the problem is that it is in the parse thread. It is visible in profiler.

* simplify `BuildFileManifestPojoizer`, make it somewhat faster (hardcode all the rules instead of storing a sequence of converters)
* move it to Starlark package, because it is not used for Python

And also, invoke `BuildFileManifestPojoizer` at rule reigster time. Previously it was called after the execution of Buck file at construction of the manifest.

Reviewed By: mzlee

shipit-source-id: 71ee1fab35ac47fe25d30d7871551495f08b78bd