Validate .bzl files on load instead on on evaluation
Summary:
Fix TODO.

`.bzl` files are meant to be validated once while loaded, not each time they are evaluated.

This is possible now because previous diff D22268617 adds tracking of load stack, which is needed for humans to better understand the error.

There are two changes in tests:
* stack trace format changed (it no longer resembles python; should not fixed, but not critical)
* more importantly, line numbers changed, because not it prints line number of load statement, not line number of function which would fail if there was no validation

(This also speeds up Starlark a little).

(Note: this ignores all push blocking failures!)

Reviewed By: mzlee

shipit-source-id: 701e0c1df5bba9c2d3d6d9d49e800aa96f148626