Add a parser for parsing args passed to StepsExecutable
Summary:
`ExternalArgsParser` is used to pass args passed to the StepsExecutable. For example, if the StepsExecutable was invoked like

  java -jar  buck-out/gen/ce9b6f2e/src/com/facebook/buck/external/main.jar foo path

`ExternalArgsParser` tries to parse the `foo path` part

`ExternalArgsParser` currently expects two args:
1. buildable class name associated with the steps (e.g. com.facebook.buck.android.SplitResources)
2. path to buildable command file

These two things will be used by another class to figure out which steps should be constructed. I.e. given the buildable name and buildable command, StepsExecutable should be able to construct the steps that need to be executed. A separate class, `StepsConstructor`, will be added to handle this.

Other variables (e.g. action ID) passed to StepsExecutable are passed through the environment, and thus will be handled separately.

Reviewed By: mykola-semko

shipit-source-id: 18c7b9fa9d12d1b561a27bf8e74724f72b7ab805