Fix Python test modules file path
Summary:
Previously, the `WriteFile` rule the Python test description
would setup used a path formed by a slightly different
build target, which caused it to conflict with the output
path used by the `#source-db` rules (and lead to the latter
deleting the former when it ran).

Have the `WriteFile` rule use an output path formed by it's
own build target to avoid this.

Reviewed By: asp2insp

shipit-source-id: fd3340d56019b1f35876072f62bf0c42d0af374e