Slightly faster rule function in Starlark
Summary:
It is not a huge perf improvement, but it is visible in profiler.

Starlark provides slower and convenient `call` API and faster "raw" `fastcall` interface.

`call` interface matches parameters to function signature and put named argments in a hash map, while `fastcall` passes arguments as as in two arrays.

This diff switches rule function implementation to `fastcall`.

Reviewed By: mzlee

shipit-source-id: 17f06564cf1e672000c17ce33f456291ce70298c