CxxLibraryFactory supports additional comp db entries
Summary: `CxxLibraryDescriptionDelegate.ConfiguredDelegate` is the interface which allows custom functionality to be injected by the rule for `apple_library`. In this diff, the interface is expanded to support the injection of additional compilation database entries. This mechanism will be used to inject entries for the Swift sources in the output of apple_library#compilation-database`.

Reviewed By: blackm00n

shipit-source-id: 30d5e05bfe11785772db6e737bfe80ed8bbf6d1a