Support Swift sources in compilation db of apple_library
Summary:
This diff connects all the pieces together so that Swift sources can appear in the compilation databases of `apple_library` provided the correct flag is turned on (i.e., `apple.compilation_database_includes_swift=true`).

It consists of the following logical changes:

### 1. Introduce `apple-swift-command` flavor

This is the new flavor for the rule that will be used to generate and write out the Swift compilation command.

### 2. Expose Rule Creation

`AppleLibraryDescriptionSwiftEnhancer.createSwiftCommandRule()` is the method that creates the rule to generate the compilation command. It will be passed a `BuildTarget` with the flavor introduced above.

Note that `AppleLibraryDescriptionSwiftEnhancer.createSwiftCommandRule()` mostly forwards `SwiftLibraryDescription.createSwiftCommandRule()` where the real work happens.

### 3. Add config option to include Swift sources

We add the `apple.compilation_database_includes_swift` configuration option which controls whether Swift sources appear in the compilation database of `apple_library`. There are two main reasons, both pertaining to new behavior:

- Performance: due to the need to compile Swift code to generate `.swiftmodule` (as `SwiftCompileBase` cannot separate compilation from module interface generation), including Swift sources in the comp database introduces a new perf cost.
- `file` Field: the `file` field includes multiple files due to whole module compilation which might break certain tooling.

### 4. Wire-up Swift Compilation DB Rules

Finally, `AppleLibraryDescription.getCompilationDatabaseRulesForDelegate()` checks whether the compilation db needs to include Swift and if so, passes the appropriate rule.

### 5. Add Integration Test

Newly added `SwiftCompilationDatabaseIntegrationTest.java` which tests the implemented functionality.

Reviewed By: blackm00n

shipit-source-id: 28a7f2b1814e46cc544c2388bea6d3884ade96b4