Move 'buck query' off the main thread
Summary:
Add buck dependency contains few steps:

1. Run buck query to find the module owner of the class you want to import (blocking)
2. Once found, add it to the BUCK file (write)
3. Add the new module to the current module dependencies so the IDE can refer to it.
4. Add Import to the file (write)

Currently, all the steps are running on the UI thread and causing the IDE to freeze until the it completed.
The first step is the most expensive once, which can run on off the main thread and not block the IDE.

The experience is much better and allows the user to keep typing - and minimize the freeze just to writing the dependency part.

2 possible scenarios after user pressed 'add buck dependency':
1. waiting without typing - in this case the command will also add the import for the class.
2. user keep typing - in this case the class element is not valid anymore since the file changed and we'll not be able to add the import, but this is super easy to add with alt+enter almost without blocking the user.

Reviewed By: cute-jumper

shipit-source-id: f1b55ce225d10f8aefec3be37c4bc6fd6d5e8073