Add args to leave extra space between secondary dexes
Summary:
This is a hack to work around discrepancies between how buck counts field references vs d8. In the long term we should use d8 itself to do dex splitting, or at least reuse the code d8 uses to count refs.

There appears to be more than one source of inaccuracy in how we are counting field references, related to static methods and enums, but they are rare and account for a small number per dex (<20).

If we fix the current bugs in ref counting, it's still possible future changes to d8 could generate more method/field references for the same code. These args can unbreak builds if that happens.

Reviewed By: IanChilds

shipit-source-id: ff0285af42960e8f3a83fa6412edac49445a71e0