Add oso_prefix support for ld64 (Darwin linker)
Summary:
# Context

Debug info for Mach-O executables can be stored either in the object files (incremental build) or in dSYMs. For the incremental use case, the executables themselves have paths to the object files which contain the DWARF debug data (the paths are contained in so called OSO entries).

ld64 inserts absolute paths in the OSO which makes the object files non-cacheable for the purposes of reproducible builds, so Buck contains a postprocessing step that relativizes the paths in the OSO entries. It has two major problems:

- It does not support binaries larger than 2GiB.
- It's slow for large binaries (1GB+) due to the need to rewrite the symtab which can be half the size of debug binaries.

# Changes

In Xcode 11, ld64 gained a new option, -oso_prefix that can strip a prefix from all OSO paths. This can be used to relativize all those paths. This diff adds support for this behind a configuration option (`cxx.link_path_normalization_args_enabled`).

Reviewed By: blackm00n

shipit-source-id: 5bcc3e689b4bcdf32880d16374679918c3704087