Fix agent version code
Summary:
The agent version code was off by one, which resulted in the agent being reinstalled every time. On my machine with a physical device, this incurred a 2-7s penalty on every `buck install` (probably less for emulators). This appears to have regressed due to my earlier D20352000 diff stack.

The apk was generated with:

```
$ buck build android/com/facebook/buck/android/agent:agent --show-output
//android/com/facebook/buck/android/agent:agent buck-out/gen/ce9b6f2e/android/com/facebook/buck/android/agent/agent.apk
$ cp buck-out/gen/ce9b6f2e/android/com/facebook/buck/android/agent/agent.apk assets/android/agent.apk
```

Reviewed By: jiawei-lyu

shipit-source-id: cb1ae5fbfbc0044b5be38059f525ffa0ab2b9459