buck/rust: Restore propagation of check builds
Summary:
When performing a Rust #check or #doc build we only need type information from the dependencies, not machine code. That corresponds to requiring an rmeta output, not an rlib.

Also use separate `crateType` and `wantCrateType` to keep the inputs and outputs distinct. Using one variable name was previously obscuring the fact that the `crateType` argument was completely unused.

Reviewed By: gtr

shipit-source-id: bb4d1447f6834b2903d626034565ac39f3f91776