apple | fix adhoc code sign performed despite dry run
Summary: Regression introduced in D22189247. Before, dry code sign flag affected ADHOC code sign type, after that change decision was based on `maybeDryCodeSignResultFile` optional which only was set for DISTRIBUTION code sign type. In this diff I revert that part of the change and provide both `maybeDryCodeSignResultFile` path and `dryRunCodeSigning` flag, the former is only copied when signing mode is DISTRIBUTION and is present, whereas the latter is actually manages if signing takes place.

shipit-source-id: 4b450c4efdce48f61b10ebad9afe0d09d333afc6