honor WATCHMAN_SOCK in WatchmanFactory
Summary:
On On-demand hosts, the effective user is twsvcscm, and everything is running as twsvcscm. However, in the Buck wrapper, Buck is masquerading to be the actual user by arbitrarily set USER environment variable. This is problematic to watchman since watchman relies on that environment variable to find path to the UNIX domain socket to communicate with the server.

However, buck does set this in order to get telemetry for users specifically, so a workaround would be to explicitly set WATCHMAN_SOCK in the environment. This diff works to honor this variable, where the previous diff sets up the environment

Reviewed By: miaoyipu

shipit-source-id: e6948526d76c501f302293e4ef595067734361b4