Fix code coverage with Java 11 by adding new config for the Java invocation
Summary:
When running tests with Buck that is built using Java 11, the invocation of Java that loads the code coverage code needs to use Java 11 else we are trying to run code that was built by a newer version of Java than we are running.

Therefore just add a new config that can point to Java 11 so that this works correctly.

Reviewed By: cute-jumper

shipit-source-id: 089aa32fb89d2b2f22d99ce0a0c977c3feabfd79