Ensure that targets in the same package are always visible AND within_view of each other
Summary:
From the [visibility documentation](https://buck.build/concept/visibility.html) (emphasis mine):

> Both attributes act as whitelists, with some exceptions. In general, if a target is not listed, there may be no dependency relationship. If the within_view list is empty or unset, however, its check is bypassed. Similarly, **targets defined in the same build file always act as if they were members of their siblings' visibility lists**.

Target's are always visible to other targets in the same package. However, as *currently implemented*, they're *not within_view* but should be. For example:

//Foo/BUCK:
```
java_library(
  name = "bar",
)

java_library(
  name = "baz",
  within_view = [
    "//qux/...",
  ],
  deps = [
    ":bar",
    "//qux:random_target1",
    "//qux:random_target2",
  ],
)
```

In this setup, `bar` is visible to `baz` but `bar` is not considered `within_view` of `baz`. I believe it should be. If targets are by default visible to targets in the package, they should also be within_view. Otherwise you'd need to list your own package or dependent targets in the `within_view`.

Fix this oversight by making targets within_view of other targets in the package.

Reviewed By: stepancheg

shipit-source-id: 0efe851e3ec6137d31d39e538e5cb76616ec7e33