apple | copy provisioning profile via AppleResourceProcessing
Summary: To further simplify AppleBundle class, there is a fine detail - provisioning profile file might not exist though all the previous steps succeeded when code sign is run in dry mode. To maintain this behavior I'm adding `boolean FileAppleBundlePart.getIgnoreIfMissing()` which is going to flag that before copying/processing the source we should check if it exist, if it doesn't exist we should handle it gracefully.

Reviewed By: milend

shipit-source-id: 83f9ada8386e99ffe9767bdaeebe3cb869ffa184