Sync with master
Summary:
* update to Starlark from master
* remove cpu profiler from Starlark
* allow recursion in Starlark
* remove certain Starlark tests which depend too much on Bazel internals
* fix Buck compilation and tests

This diff is not very reviewable.

They changed quite a lot since the last sync:
* renamed a lot of classes and method
* changed signatures of functions like `fastcall` (which no longer accepts `Location` for example)
* shuffled a lot of utilities
* strict bzl validation is now mandatory (which affected a lot of our tests)
* bazel-specific functions (like `select` or `depset`) are not longer a part of Starlark core
* and so on

I don't know how I could split it into smaller pieces, so maybe accept it as is.

Bazel accepted certain optimizations, for example:
* [optimize list construction](https://github.com/bazelbuild/bazel/pull/11787)
* [optimize dict construction](https://github.com/bazelbuild/bazel/pull/11785)
* [faster reflection cache](https://github.com/bazelbuild/bazel/pull/11681)

Also, I see that Starlark is slowly moving into a standalone interpreter (they reduce the dependencies on Bazel; for example `select` is no longer a part of Starlark).

So might be reasonable to continue synching with bazel repository.

Reviewed By: scottrice

shipit-source-id: 5ad06d72daa3c1bde994d618139ff76c046182d9